Split deposit plugins: index.ts as barrel, plugin.ts holds logic
Each deposit method now has index.ts (re-exports only) and plugin.ts
(the actual register function + PaymentsPlugin implementation). Approval
watcher for pull-deposit remains in its own file.
